FAD-Brain/FAD-Brain/requirements.txt
# -----------------------------
# Core Networking (REQUIRED)
# -----------------------------
pynetworktables>=2021.0.0
zeromq>=0.0.0 ; extra == "optional"
pyzmq>=25.0.0 ; extra == "optional"

# -----------------------------
# Vision / Video
# -----------------------------
opencv-python>=4.8.0
numpy>=1.24.0

# -----------------------------
# ML / Inference (CPU-safe)
# -----------------------------
torch>=2.1.0
torchvision>=0.16.0
onnx>=1.15.0
onnxruntime>=1.17.0

# -----------------------------
# Performance / Concurrency
# -----------------------------
numba>=0.58.0
psutil>=5.9.0

# -----------------------------
# Data / Config
# -----------------------------
pyyaml>=6.0.1
dataclasses-json>=0.6.4

# -----------------------------
# Logging / Debug
# -----------------------------
rich>=13.7.0
loguru>=0.7.2

# -----------------------------
# Optional: Web / APIs (DS dashboards)
# -----------------------------
fastapi>=0.110.0
uvicorn>=0.27.0

# -----------------------------
# Optional: Dev / Testing
# ---------
